requirements.txt
pandas
numpy<2.0
scikit-learn
scikit-surprise
reportlab